+++ node/src/Worker.ts
@@ -22,7 +22,7 @@ import { WebRtcServerImpl } from './WebRtcServer';
 import type { Router, RouterOptions } from './RouterTypes';
 import { RouterImpl } from './Router';
 import { portRangeToFbs, socketFlagsToFbs } from './Transport';
-import type { RtpCodecCapability } from './rtpParametersTypes';
+import type { RouterRtpCodecCapability } from './rtpParametersTypes';
 import * as utils from './utils';
 import * as fbsUtils from './fbsUtils';
 import type { AppData } from './types';
@@ -533,9 +533,9 @@ export class WorkerImpl<WorkerAppData extends AppData = AppData>
 		}
 
 		// Clone given media codecs to not modify input data.
-		const clonedMediaCodecs = utils.clone<RtpCodecCapability[] | undefined>(
-			mediaCodecs
-		);
+		const clonedMediaCodecs = utils.clone<
+			RouterRtpCodecCapability[] | undefined
+		>(mediaCodecs);
 
 		// This may throw.
 		const rtpCapabilities =